+++ src/pyatmo/modules/__init__.py
@@ -1,10 +1,14 @@
 """Expose submodules."""
 from .base_class import Place
-from .bticino import BNCX, BNDL, BNSL
+from .bticino import BNCX, BNDL, BNEU, BNSL
 from .idiamant import NBG, NBO, NBR, NBS
 from .legrand import (
+    EBU,
+    NLAO,
+    NLAS,
     NLC,
     NLD,
+    NLDD,
     NLE,
     NLF,
     NLFE,
@@ -12,6 +16,7 @@ from .legrand import (
     NLG,
     NLIS,
     NLL,
+    NLLF,
     NLLM,
     NLLV,
     NLM,
@@ -23,9 +28,13 @@ from .legrand import (
     NLPS,
     NLPT,
     NLT,
+    NLTS,
     NLUF,
     NLUI,
+    NLUO,
+    NLUP,
     NLV,
+    Z3L,
     NLunknown,
 )
 from .module import Camera, Dimmer, Module, Shutter, Switch
@@ -52,10 +61,12 @@ from .netatmo import (
     PublicWeatherArea,
 )
 from .smarther import BNS
+from .somfy import TPSRS
 
 __all__ = [
     "BNCX",
     "BNDL",
+    "BNEU",
     "BNS",
     "BNSL",
     "Camera",
@@ -81,6 +92,7 @@ __all__ = [
     "NIS",
     "NLC",
     "NLD",
+    "NLDD",
     "NLE",
     "NLF",
     "NLFE",
@@ -101,8 +113,14 @@ __all__ = [
     "NLT",
     "NLUF",
     "NLUI",
+    "NLAO",
+    "NLLF",
+    "NLUO",
     "NLunknown",
+    "NLUP",
     "NLV",
+    "EBU",
+    "Z3L",
     "NOC",
     "NRV",
     "NSD",
@@ -112,4 +130,7 @@ __all__ = [
     "PublicWeatherArea",
     "Shutter",
     "Switch",
+    "TPSRS",
+    "NLAS",
+    "NLTS",
 ]