+++ src/pyatmo/modules/device_types.py
@@ -42,12 +42,10 @@ class DeviceType(str, Enum):
     # Home Coach
     NHC = "NHC"  # Smart Indoor Air Quality Monitor
 
-    # 3rd Party
-    BNS = "BNS"  # Smarther with Netatmo
-
     # Legrand Wiring devices and electrical panel products
     NLC = "NLC"  # Cable outlet
     NLD = "NLD"  # Dimmer
+    NLDD = "NLDD"  # Dimmer
     NLE = "NLE"  # Connected Ecometer
     NLF = "NLF"  # 2 wire light switch
     NLFN = "NLFN"  # light switch with neutral
@@ -67,9 +65,15 @@ class DeviceType(str, Enum):
     NLPT = "NLPT"  # Connected latching relay / Telerupt
     NLT = "NLT"  # Global remote control
     NLV = "NLV"  # Legrand / BTicino shutters
-    NLUI = "NLUI"  # Legrand device stub
+    NLAO = "NLAO"  # Legrand wireless batteryless light switch
+    NLUO = "NLUO"  # Legrand Plug-In dimmer switch
+    NLUI = "NLUI"  # Legrand In-Wall ON/OFF switch
     NLunknown = "NLunknown"  # Legrand device stub
     NLUF = "NLUF"  # Legrand device stub
+    NLAS = "NLAS"  # Legrand wireless batteryless scene switch
+    NLUP = "NLUP"  # Legrand device stub
+    NLLF = "NLLF"  # Legrand device stub
+    NLTS = "NLTS"  # Legrand motion sensor stub
 
     # BTicino Classe 300 EOS
     BNCX = "BNCX"  # internal panel = gateway
@@ -82,6 +86,14 @@ class DeviceType(str, Enum):
     NBO = "NBO"  # orientable shutter
     NBR = "NBR"  # roller shutter
     NBS = "NBS"  # swing shutter
+
+    # Somfy
+    TPSRS = "TPSRS"  # Somfy io shutter
+
+    # 3rd Party
+    BNS = "BNS"  # Smarther with Netatmo
+    EBU = "EBU"  # EBU gas meter
+    Z3L = "Z3L"  # Zigbee 3 Light
 
     # pylint: enable=C0103
 
@@ -124,27 +136,37 @@ DEVICE_CATEGORY_MAP: dict[DeviceType, DeviceCategory] = {
     DeviceType.NLLV: DeviceCategory.shutter,
     DeviceType.NLLM: DeviceCategory.shutter,
     DeviceType.NBR: DeviceCategory.shutter,
+    DeviceType.NBO: DeviceCategory.shutter,
     DeviceType.NLP: DeviceCategory.switch,
     DeviceType.NLPM: DeviceCategory.switch,
     DeviceType.NLPBS: DeviceCategory.switch,
     DeviceType.NLIS: DeviceCategory.switch,
     DeviceType.NLL: DeviceCategory.switch,
+    DeviceType.NLM: DeviceCategory.switch,
+    DeviceType.NLC: DeviceCategory.switch,
     DeviceType.NLFN: DeviceCategory.dimmer,
     DeviceType.NLF: DeviceCategory.dimmer,
     DeviceType.BNS: DeviceCategory.climate,
     DeviceType.NLPC: DeviceCategory.meter,
     DeviceType.NLE: DeviceCategory.meter,
+    DeviceType.Z3L: DeviceCategory.dimmer,
+    DeviceType.NLUP: DeviceCategory.switch,
+    DeviceType.NLPO: DeviceCategory.switch,
+    DeviceType.TPSRS: DeviceCategory.shutter,
+    DeviceType.NLUO: DeviceCategory.dimmer,
+    DeviceType.NLUI: DeviceCategory.switch,
+    DeviceType.NLUF: DeviceCategory.dimmer,
 }
 
 
 DEVICE_DESCRIPTION_MAP: dict[DeviceType, tuple[str, str]] = {
-    # Climate/Energy
+    # Netatmo Climate/Energy
     DeviceType.NAPlug: ("Netatmo", "Smart Thermostat Gateway"),
     DeviceType.NATherm1: ("Netatmo", "Smart Thermostat"),
     DeviceType.NRV: ("Netatmo", "Smart Valve"),
     DeviceType.OTH: ("Netatmo", "OpenTherm Gateway"),
     DeviceType.OTM: ("Netatmo", "OpenTherm Modulating Thermostat"),
-    # Cameras/Security,
+    # Netatmo Cameras/Security
     DeviceType.NOC: ("Netatmo", "Smart Outdoor Camera"),
     DeviceType.NACamera: ("Netatmo", "Smart Indoor Camera"),
     DeviceType.NSD: ("Netatmo", "Smart Smoke Detector"),
@@ -152,18 +174,16 @@ DEVICE_DESCRIPTION_MAP: dict[DeviceType, tuple[str, str]] = {
     DeviceType.NACamDoorTag: ("Netatmo", "Smart Door/Window Sensors"),
     DeviceType.NDB: ("Netatmo", "Smart Video Doorbell"),
     DeviceType.NCO: ("Netatmo", "Smart Carbon Monoxide Alarm"),
-    # Weather,
+    # Netatmo Weather
     DeviceType.NAMain: ("Netatmo", "Smart Home Weather station"),
     DeviceType.NAModule1: ("Netatmo", "Smart Outdoor Module"),
     DeviceType.NAModule2: ("Netatmo", "Smart Anemometer"),
     DeviceType.NAModule3: ("Netatmo", "Smart Rain Gauge"),
     DeviceType.NAModule4: ("Netatmo", "Smart Indoor Module"),
     DeviceType.public: ("Netatmo", "Public Weather station"),
-    # Home Coach,
+    # Netatmo Home Coach
     DeviceType.NHC: ("Netatmo", "Smart Indoor Air Quality Monitor"),
-    # 3rd Party,
-    DeviceType.BNS: ("Smarther", "Smarther with Netatmo"),
-    # Legrand Wiring devices and electrical panel products,
+    # Legrand Wiring devices and electrical panel products
     DeviceType.NLG: ("Legrand", "Gateway"),
     DeviceType.NLGS: ("Legrand", "Gateway standalone"),
     DeviceType.NLP: ("Legrand", "Plug"),
@@ -184,14 +204,28 @@ DEVICE_DESCRIPTION_MAP: dict[DeviceType, tuple[str, str]] = {
     DeviceType.NLPS: ("Legrand", "Smart Load Shedder"),
     DeviceType.NLC: ("Legrand", "Cable Outlet"),
     DeviceType.NLT: ("Legrand", "Global Remote Control"),
-    # BTicino Classe 300 EOS,
+    DeviceType.NLAS: ("Legrand", "Wireless batteryless scene switch"),
+    DeviceType.NLD: ("Legrand", "Dimmer"),
+    DeviceType.NLDD: ("Legrand", "Dimmer"),
+    DeviceType.NLUP: ("Legrand", "Power outlet"),
+    DeviceType.NLUO: ("Legrand", "Plug-In dimmer switch"),
+    DeviceType.NLUI: ("Legrand", "In-wall switch"),
+    DeviceType.NLTS: ("Legrand", "Motion sensor"),
+    DeviceType.NLUF: ("Legrand", "In-Wall dimmer"),
+    # BTicino Classe 300 EOS
     DeviceType.BNCX: ("BTicino", "Internal Panel"),
     DeviceType.BNEU: ("BTicino", "External Unit"),
     DeviceType.BNDL: ("BTicino", "Door Lock"),
     DeviceType.BNSL: ("BTicino", "Staircase Light"),
-    # Bubbendorf shutters,
+    # Bubbendorf shutters
     DeviceType.NBG: ("Bubbendorf", "Gateway"),
     DeviceType.NBR: ("Bubbendorf", "Roller Shutter"),
     DeviceType.NBO: ("Bubbendorf", "Orientable Shutter"),
     DeviceType.NBS: ("Bubbendorf", "Swing Shutter"),
+    # Somfy
+    DeviceType.TPSRS: ("Somfy", "io Shutter"),
+    # 3rd Party
+    DeviceType.BNS: ("Smarther", "Smarther with Netatmo"),
+    DeviceType.Z3L: ("3rd Party", "Zigbee 3 Light"),
+    DeviceType.EBU: ("3rd Party", "EBU gas meter"),
 }