+++ src/pyatmo/room.py
@@ -67,12 +67,12 @@ class Room(NetatmoBase):
             if module.device_category is not None:
                 self.features.add(module.device_category.name)
 
-        if "NRV" in self.device_types:
+        if "OTM" in self.device_types:
+            self.climate_type = DeviceType.OTM
+        elif "NRV" in self.device_types:
             self.climate_type = DeviceType.NRV
         elif "NATherm1" in self.device_types:
             self.climate_type = DeviceType.NATherm1
-        elif "OTM" in self.device_types:
-            self.climate_type = DeviceType.OTM
         elif "BNS" in self.device_types:
             self.climate_type = DeviceType.BNS
             self.features.add("humidity")
@@ -107,7 +107,9 @@ class Room(NetatmoBase):
         """Set room temperature set point."""
         mode = MODE_MAP.get(mode, mode)
 
-        if "NATherm1" in self.device_types or "NRV" in self.device_types:
+        if "NATherm1" in self.device_types or (
+            "NRV" in self.device_types and not self.home.has_otm()
+        ):
             await self._async_set_thermpoint(mode, temp, end_time)
 
         else: